type

+++ src/enc.c
@@ -4,7 +4,7 @@ void cipher(struct gen *g, uint8_t *buf, size_t len)
 {
         uint32_t stream_pos;
         size_t ncalls = len >> 2;
-        uint8_t left = len & 3;
+        size_t left = len & 3;
         uint8_t *tmp = (uint8_t *)&stream_pos;
 
         for(size_t i = 0; i < ncalls; ++i) {
@@ -16,7 +16,7 @@ void cipher(struct gen *g, uint8_t *buf, size_t len)
         }
         if(left != 0) {
                 stream_pos = gen32(g);
-                for(uint8_t i = 0; i < left; ++i)
+                for(size_t i = 0; i < left; ++i)
                         buf[(ncalls<<2)+i] ^= tmp[i];
         }
 }
